+++ projects/igniteui-angular/src/lib/query-builder/query-builder-header.component.ts
@@ -36,6 +36,8 @@ export class IgxQueryBuilderHeaderComponent {
     /**
      * Sets the resource strings.
      * By default it uses EN resources.
+     * 
+     * @deprecated in version 19.1.0.
      */
     @Input()
     public set resourceStrings(value: IQueryBuilderResourceStrings) {